firewalld: set public zone target to REJECT
Bring forward the forthcoming change from firewalld 1.0:
https://github.com/firewalld/firewalld/commit/f2896e43c3a548a299f87675a01e1a421b8897b8

After that commit lands, 'default' is the same as 'REJECT' except that
it accepts ICMP by default. It is not yet clear whether it will forward
ICMP by default.

Until then, let's not use 'default' at all. A rich rule is used to
allow the host to be pinged, however.