Configure chrony on RHEL
The same mechanism will be used on Debian when I get around to it.

We're doing it in common.yml because we can't apparantly have a task in
the os-specific tasks file notify a handler in the playbook that
imported it.

So we may as well move OS-specific variables out to separate variables
files, and use them to customize the tasks within the common playbook,
as described in
<https://ansible-tips-and-tricks.readthedocs.io/en/latest/os-dependent-tasks/variables/>.

Interestingly
<https://ansible-tips-and-tricks.readthedocs.io/en/latest/os-dependent-tasks/variables/#task-separation-via-includes-deprecated-in-20>
says that having per-OS tasks via includes is deprecated for some time
now. Maybe it would be better to move everything into common.yml... but
there's rather a lot that is very OS-family specific, such as enabling
repositories. So I think I'll keep the separate tasks files for the time
being, but maybe move what I can back out to common.yml.