ipa-server: abandon persistent searches for crl fetch script
Persistent searches seem to work fine when watching a user's entry, but
they never give me any results when watching CRLs. Also it's not clear
how to handle the server restarting; and the behaviour observed when
searching with changes_only=False seems to contractict the internet
draft (the initial search completes and the persistent search never
provides any changes, necessetating a normal search + a persitant search
as well). Oh and the ldap3 abstraction layer is quite nice, but it's not
clear how to use it with persistent searches.

So now we do a regular search, and schedule the next search for (more or
less) half way between now and the earliest expiry time of the CRLs
found.